+++ src/main/ng/src/app/main/content/pages/applicationFee/inceptionFeeReceipt/inceptionFeeReceipt.component.ts
@@ -101,6 +101,9 @@ export class InceptionFeeReceiptComponent {
         // Subscribe to the dialog close event to intercept the action taken.
         dialogRef.afterClosed().subscribe((result) => {
             if (result.refresh) {
+                this._applicationFeeService.getApplicationFee(this.loanApplicationId).subscribe(data => {
+                    this._applicationFeeService._applicationFee.next(data);
+                });
                 this.refreshTable();
             }
         });
@@ -112,10 +115,13 @@ export class InceptionFeeReceiptComponent {
     delete(): void {
         const dialogRef = this._matDialog.open(ConfirmationDialogComponent);
         // Subscribe to the dialog close event to intercept the action taken.
-        dialogRef.afterClosed().subscribe((response) => {
-            if (response) {
+        dialogRef.afterClosed().subscribe((result) => {
+            if (result && result.response) {
                 this._applicationFeeService.deleteInceptionFee(this.selectedInceptionFee.id).subscribe(() => {
                     this.selectedInceptionFee = undefined;
+                    this._applicationFeeService.getApplicationFee(this.loanApplicationId).subscribe(data => {
+                        this._applicationFeeService._applicationFee.next(data);
+                    });
                     this.refreshTable();
                 },
                 (error) => {
