+++ src/main/ng/src/app/main/content/pages/iccApproval/loanEnhancement/loanEnhancement.component.ts
@@ -94,9 +94,12 @@ export class LoanEnhancementComponent implements OnInit {
     deleteLoanEnhancement(): void {
         const dialogRef = this._dialog.open(ConfirmationDialogComponent);
         // Subscribe to the dialog close event to intercept the action taken.
-        dialogRef.afterClosed().subscribe((response) => {
-            if (response) {
+        dialogRef.afterClosed().subscribe((result) => {
+            if (result && result.response) {
                 this._iccApprovalService.deleteLoanEnhancement(this.selectedLoanEnhancement.id).subscribe(() => {
+                    this._iccApprovalService.getICCApproval(this.loanApplicationId).subscribe(data => {
+                        this._iccApprovalService._iccApproval.next(data);
+                    });
                     this.selectedLoanEnhancement = undefined;
                     this.refreshTable();
                 });
